Refactor buildOptions into own file
- this file is way too long
- the helper is also used by other files
